+++ packages/appinio_video_player/lib/appinio_video_player.dart
@@ -1,6 +1,6 @@
 library appinio_video_player;
 
-export 'package:video_player/video_player.dart';
+export 'package:cached_video_player/cached_video_player.dart';
 export 'src/custom_video_player.dart';
 export 'src/web_video_player/conditional_native_web_video_player/conditional_native_web_video_player.dart';
 export 'src/models/custom_video_player_progress_bar_settings.dart';