+++ packages/widget_zoom/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.0.4] - 2024.03.07
+
+- Added `closeFullScreenImageOnDispose` option to `WidgetZoom`.
+
+## [0.0.3] - 2023.12.21
+
+- Updated readme.
+
+
 ## [0.0.2] - 2023.10.16
 
 - Updated readme.